Refactor code move everything to template class allow user to configure layout option
